requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.3
PyJWT==2.5.0
python-dotenv==0.21.0
python-http-client==3.3.7
pytz==2022.2.1
requests==2.28.1
sendgrid==6.9.7
starkbank-ecdsa==2.1.0
twilio==7.14.1
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
